Rename CacheComponent to CacheControllerComponent, etc.
